+++ lib/proxy/proxy.dart
@@ -1,5 +1,6 @@
 export 'axis.dart';
 export 'button.dart';
+export 'image.dart';
 export 'listTile.dart';
 export 'radio.dart';
 export 'textFormField.dart';